Mediatype icons saved in DetailledItem and shown in starred list (fixes #321)

+++ opacclient/opacapp/src/main/java/de/geeksfactory/opacclient/storage/StarDatabase.java
@@ -31,16 +31,16 @@ public class StarDatabase extends SQLiteOpenHelper {
     public static final String STAR_TABLE = "starred";
     private static final String DATABASE_CREATE = "create table " + STAR_TABLE
             + " ( id integer primary key autoincrement," + " medianr text,"
-            + " bib text," + " title text" + ");";
+            + " bib text," + " title text," + " mediatype text" + ");";
     // CHANGE THIS
     public static final String STAR_WHERE_ID = "id = ?";
     public static final String STAR_WHERE_LIB = "bib = ?";
     public static final String STAR_WHERE_TITLE_LIB = "bib = ? AND title = ?";
     public static final String STAR_WHERE_NR_LIB = "bib = ? AND medianr = ?";
     public static final String[] COLUMNS = {"id AS _id", "medianr", "bib",
-            "title"};
+            "title", "mediatype"};
     private static final String DATABASE_NAME = "starred.db";
-    private static final int DATABASE_VERSION = 5; // REPLACE ONUPGRADE IF YOU
+    private static final int DATABASE_VERSION = 6; // REPLACE ONUPGRADE IF YOU
 
     public StarDatabase(Context context) {
         super(context, DATABASE_NAME, null, DATABASE_VERSION);
@@ -53,19 +53,26 @@ public class StarDatabase extends SQLiteOpenHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        Log.w(StarDatabase.class.getName(), "Upgrading database from version "
-                + oldVersion + " to " + newVersion
-                + ", which will destroy all old data");
 
-        db.execSQL(
-                "create table temp ( id integer primary key autoincrement, medianr text, " +
-                        "bib text, title text );");
-        db.execSQL("insert into temp select * from starred;");
-        db.execSQL("drop table starred;");
-        onCreate(db);
-        db.execSQL("insert into starred select * from temp;");
-        db.execSQL("drop table temp;");
+        if (oldVersion >= 5) {
+            if (oldVersion < 6) {
+                // Add column for media type
+                db.execSQL("alter table " + STAR_TABLE + " add column mediatype text");
+            }
+        } else {
+            Log.w(StarDatabase.class.getName(), "Upgrading database from version "
+                    + oldVersion + " to " + newVersion
+                    + ", which will destroy all old data");
 
+            db.execSQL(
+                    "create table temp ( id integer primary key autoincrement, medianr text, " +
+                            "bib text, title text );");
+            db.execSQL("insert into temp select * from starred;");
+            db.execSQL("drop table starred;");
+            onCreate(db);
+            db.execSQL("insert into starred select * from temp;");
+            db.execSQL("drop table temp;");
+        }
     }
 
 }
